Add files via upload
The project starts with a set of business requirements. The business problem we are going to address is providing a report of upcoming projects for a general contracting business.

A general contractor is hired for various type of projects from building a house to minor repairs. For this application, the contractor wants to keep track of the employees working on a project, the project customer’s name and address, and the start and end date for the project.